requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.5.0
numpy>=1.24.0
requests>=2.31.0
fairseq
timm
natsort
tensorboard
nltk
h5py
fastwer
git+https://github.com/pytorch/fairseq.git
